+++ oci/layer/types_test.go
@@ -19,6 +19,7 @@
 package layer
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -132,4 +133,41 @@ func TestRepackOptionsHelpers(t *testing.T) {
 
 		assert.Equal(t, onDiskFmt.MapOptions, overlayOpt.MapOptions(), "MapOptions should match actual MapOptions")
 	})
+}
+
+func TestOverlayfsRootfs_XattrNamespace(t *testing.T) {
+	for _, test := range []struct {
+		onDiskFmt         OverlayfsRootfs
+		expectedNamespace string
+	}{
+		{OverlayfsRootfs{UserXattr: false}, "trusted."},
+		{OverlayfsRootfs{UserXattr: true}, "user."},
+	} {
+		test := test // copy iterator
+		t.Run(fmt.Sprintf("UserXattr=%v", test.onDiskFmt.UserXattr), func(t *testing.T) {
+			assert.Equalf(t, test.expectedNamespace, test.onDiskFmt.xattrNamespace(), "OverlayfsRootfs{UserXattr: %v}.xattrNamespace()", test.onDiskFmt.UserXattr)
+		})
+	}
+}
+
+func TestOverlayfsRootfs_Xattr(t *testing.T) {
+	for _, test := range []struct {
+		name      string
+		onDiskFmt OverlayfsRootfs
+		subxattr  []string
+		expected  string
+	}{
+		{"TrustedXattr-NoArgs", OverlayfsRootfs{UserXattr: false}, []string{}, "trusted.overlay."},
+		{"UserXattr-NoArgs", OverlayfsRootfs{UserXattr: true}, []string{}, "user.overlay."},
+		{"TrustedXattr-SubXattr", OverlayfsRootfs{UserXattr: false}, []string{"opaque"}, "trusted.overlay.opaque"},
+		{"UserXattr-SubXattr", OverlayfsRootfs{UserXattr: true}, []string{"whiteout"}, "user.overlay.whiteout"},
+		{"TrustedXattr-MultiSubXattr", OverlayfsRootfs{UserXattr: false}, []string{"foo", "bar.baz", "boop"}, "trusted.overlay.foo.bar.baz.boop"},
+		{"UserXattr-MultiSubXattr", OverlayfsRootfs{UserXattr: true}, []string{"abc", "def", "ghi.jkl."}, "user.overlay.abc.def.ghi.jkl."},
+	} {
+		test := test // copy iterator
+		t.Run(test.name, func(t *testing.T) {
+			got := test.onDiskFmt.xattr(test.subxattr...)
+			assert.Equalf(t, test.expected, got, "OverlayfsRootfs{UserXattr: %v}.xattr(%#v)", test.onDiskFmt.UserXattr, test.subxattr)
+		})
+	}
 }