+++ oci/layer/xattr_test.go
@@ -61,14 +61,18 @@ func TestGetXattrFilter(t *testing.T) {
 		{"system.nfs4_foo", nil, false},
 		{"system.foo", nil, false},
 		// prefixes
-		{"trusted.overlay.opaque", overlayXattrFilter{}, true},
-		{"trusted.overlay.whiteout", overlayXattrFilter{}, true},
-		{"trusted.overlay.foobar", overlayXattrFilter{}, true},
-		{"trusted.overlay.a.b.c", overlayXattrFilter{}, true},
+		{"trusted.overlay.opaque", overlayXattrFilter{"trusted."}, true},
+		{"trusted.overlay.whiteout", overlayXattrFilter{"trusted."}, true},
+		{"trusted.overlay.foobar", overlayXattrFilter{"trusted."}, true},
+		{"trusted.overlay.a.b.c", overlayXattrFilter{"trusted."}, true},
 		{"trusted.overlay", nil, false},
+		{"user.overlay.opaque", overlayXattrFilter{"user."}, true},
+		{"user.overlay.whiteout", overlayXattrFilter{"user."}, true},
+		{"user.overlay.foobar", overlayXattrFilter{"user."}, true},
+		{"user.overlay.a.b.c", overlayXattrFilter{"user."}, true},
+		{"user.overlay", nil, false},
 		// unrelated
 		{"user.foo.bar", nil, false},
-		{"user.overlay.opaque", nil, false},
 		{"user.trusted.overlay.opaque", nil, false},
 		{"user.rootlesscontainers", nil, false},
 	} {
@@ -86,11 +90,16 @@ func TestOverlayXattrFilter(t *testing.T) {
 		toDisk, toTar string
 	}{
 		{"NormalXattr", "trusted.example.xattr", OverlayfsRootfs{}, "trusted.example.xattr", "trusted.example.xattr"},
+		// With UserXattr == false, only trusted.overlay.* xattrs are affected by the filter.
 		{"TrustedOverlayXattr", "trusted.overlay.foo", OverlayfsRootfs{}, "trusted.overlay.overlay.foo", ""},
 		{"TrustedOverlayXattr-Escaped", "trusted.overlay.overlay.foo", OverlayfsRootfs{}, "trusted.overlay.overlay.overlay.foo", "trusted.overlay.foo"},
-		// TODO: Implement support for these.
 		{"UserOverlayXattr", "user.overlay.foo", OverlayfsRootfs{}, "user.overlay.foo", "user.overlay.foo"},
 		{"UserOverlayXattr-Escaped", "user.overlay.overlay.foo", OverlayfsRootfs{}, "user.overlay.overlay.foo", "user.overlay.overlay.foo"},
+		// But with UserXattr == true, only user.overlay.* xattrs are affected by the filter.
+		{"UserXattr-TrustedOverlayXattr", "trusted.overlay.foo", OverlayfsRootfs{UserXattr: true}, "trusted.overlay.foo", "trusted.overlay.foo"},
+		{"UserXattr-TrustedOverlayXattr-Escaped", "trusted.overlay.overlay.foo", OverlayfsRootfs{UserXattr: true}, "trusted.overlay.overlay.foo", "trusted.overlay.overlay.foo"},
+		{"UserXattr-UserOverlayXattr", "user.overlay.foo", OverlayfsRootfs{UserXattr: true}, "user.overlay.overlay.foo", ""},
+		{"UserXattr-UserOverlayXattr-Escaped", "user.overlay.overlay.foo", OverlayfsRootfs{UserXattr: true}, "user.overlay.overlay.overlay.foo", "user.overlay.foo"},
 	} {
 		test := test // copy iterator
 		t.Run(test.name, func(t *testing.T) {
@@ -98,7 +107,7 @@ func TestOverlayXattrFilter(t *testing.T) {
 			if !ok {
 				// For test purposes, use a dummy overlayXattrFilter if the
 				// xattr is not the right xattr.
-				filter = overlayXattrFilter{}
+				filter = overlayXattrFilter{"trusted."}
 			}
 
 			expectMasked := test.toTar == ""